+++ tests/test_kdotp.py
@@ -22,8 +22,14 @@ def test_kdotp_model():
     assert np.allclose(model.hamilton((0, 0)), np.eye(2))
     assert np.allclose(model.hamilton((1, 0)), [[1, 1j], [-1j, 1]])
     assert np.allclose(model.hamilton((0, 0.5)), [[1, 0.25], [0.25, 1]])
+    assert np.allclose(
+        model.hamilton([(0, 0), (0, 0.5), (1, 0)]),
+        [np.eye(2), [[1, 0.25], [0.25, 1]], [[1, 1j], [-1j, 1]]]
+    )
 
     assert np.allclose(model.eigenval((0, 0)), [1, 1])
+    assert np.allclose(model.eigenval([(0, 0), (0, 0)]), [[1, 1], [1, 1]])
+    assert np.allclose(model.eigenval([(0, 0), (0, 0), (0, 0.5)]), [[1, 1], [1, 1], [0.75, 1.25]])
 
 
 def test_raises_not_hermitian():